requirements.txt
tensorflow==2.17.0
numpy==1.26.4
Pillow==10.4.0
streamlit==1.32.0
protobuf==4.25.3
grpcio==1.62.2
